requirements.txt
django-allauth
django==2.0
django-rest-framework
django-tinymce
reportlab
coreapi
